A commitment to digital technologies, e.g. AI, for engineering education and learning management and productivity.
Throughout documented history, technology has shaped the art and science of engineering education and learning management in profound ways. As an example, engineering education and learning productivity itself has moved from the analog simulation of teaching and learning experiences (analog simulation of landing on the moon) through digital simulation of learning nuggets (for instance the use of applets to teach Mathematics using the Chinese Abacus, the learner-content dynamic interactivity brought forth by computer languages like Wofram Language in Mathematica etc)  to the current extended reality which will be the order of any engineering teaching and learning world in unfolding days ahead.

Hence tracking the impact of technologies like AI on engineering education and learning ought not  be left to the enthusiastic AI developer or the apprehensive politician, faculty deans but a matter for all have a stake in the management and productivity in engineering education and learning. 

This is what drives my development journey with a singular focus of becoming a hero from zero.